[WIP] DEX Activation & Snapshotting methods
- Adds a property `_dex_on` to the DEX that can be controlled
by the governance address
- Removes the old `snapshot_id`s that monotonically increase
- Switches to a day-based timekeeping system to match loans
- Add a `_check_distributions` and `_take_new_day_snapshot` method

This update is to allow each transaction on the dex to sync
with the dividends and rewards contracts if necessary. Adding
this partial commit to allow collaboration more easily, but it
should not be trusted as stable yet.

NOTE: `U_SECONDS_DAY` is defined as 1/1000th of a day for
easier testing in this build.