[DEX] Withdraw UX + Remove old slippage
The old slippage computation method wasn't effective. This removes
it so that a new one can be dropped in. Additionally, this improves
`remove` liquidity UX.

The user can set the switch to `False` (default) to just perform a
remove, keeping the balance in the contract. The user can send
`True` when doing a withdraw operation to send transactions directly
to their wallet. The contract flow is better if switching liquidity
between pools, while the all-in-one flow is better if the user
wishes to arbitrage or shift their position elsewhere.