Data sources and DEX Slippage
Adds back in the maximum slippage to the DEX, adds some bufixes to invalid pool IDs, and
links newly created data sources to the rewards score. DEX supports a maximum slippage,
specified in basis points as an integer (param `maxSlippage`). This is default to 250
(2.5%), and can be any value between 0 and 10%. If a trade would result in greater than
the maximum slippage after fees, the transaction is instead reverted.

Additionally, all named rewarded pools are now linked into the rewards SCORE by calling
the add data source API. This reduces amount of manual work in setting markets and room
for error.

- Dex has a maximum slippage (reintroduced after old implementation removed)
- Some bugfixes are added to the jupyter and smart contract APIs that would get an invalid
pool id
- New Jupyter notebook call added for the trading the reverse into the pool, so both
buys and sells can be demonstrated