Add range checks to `getDataBatch`.
Updates the function `getDataBatch()` so that the caller cannot go out of bounds.
If the caller would go out of bounds, the limit + offset are adjusted first to fit the offset within range, then to fit the limit within range of the total.

In a private node test, this stopped the dex from crashing as new transactions are sent in.