[DEX] Cleanup & Update to prices logic within the dex
The DEX formerly used the terminology of price/inverse price. This was confusing,
error-prone, and not readily easy to follow. Now, the following is in price:

- getBasePriceInQuote - Gets base currency, priced in quote currency
- getQuotePriceInBase - Gets quote currency, priced in base currency
- getPrice - alias of get base price in quote (e.g. BTC/USDT, for a common price symbol)

This gives a more natural way to lookup and access prices. Additionally, the following
other changes have been made:

- Switch math system to use EXA instead of 10**10 for a scale factor. This is used
in the loans code, and in the Band oracle, so is the accepted way to handle this data.
- Adds a reference to the BALN token to the Dex, so we can add specific methods
referencing it
- Adds a getBalnPrice, which is an alias to the BALNICD price. This will be used
to judge pool APR.