[DEX] Make basecoin flexible to a max
Before, the pool expected an exact amount to be supplied. This was difficult, as the user
had to supply amounts in the exact proportion to the pool. Now, the pool will use an exact
amount of quote coin, and up to the needed amount of basecoin to fit the pool requirement.

For example, a pool might have a 2 - 1 ratio of basecoin to quote coin. A use supplying 1 ICD could send
2, or 2 + some small buffer of basecoin and a proportional amount to the quotecoin would be supplied.
Unused basecoin is left on deposit.