[DEX] Replace withdraw lock with 24h avg
Instead of a withdrawal lock, use a 24 hour average on supplied liquidity.

Now users that supply liquidity the minute before, get only 1 minute worth of supplied liquidity time. This improves UX around a withdrawal lock, and always gives users money when they need it.

This change is implemented by storing a day long average, time of last change, and last value in all of the snapshot history maps. When getting one of these values, users should extract a weighted average (with average weighted against the start of the day, and current weighted against the remaining time in the day). This is a breaking change, and will require a new contract deployment.