[DEX] Governance set quote coins #33
This updates both the DEX and governance score to allow adding alternative quote currencies. In the short term, this will be used by projects partnering with balnaced, such as OMM. In the long term, the community will vote on which coins can be on the right side in the governance score.

sICX and bnUSD are automatically added as the tokens are configured in the DEX.

**Calling this function**

A sample cell to paste in to allow `baln` as a quote currency (switch with your choice of coin) is:

```
transaction = CallTransactionBuilder()\
  .from_(btest_wallet.get_address())\
	.to(contracts['governance']['SCORE'])\
	.value(0)\
	.step_limit(10000000)\
	.nid(NID)\
  .nonce(100)\
	.method("dexAddQuoteCoin")\
  .params({'_address': contracts['baln']['SCORE']}) \
  .build()

signed_transaction = SignedTransaction(transaction, btest_wallet)
result = icon_service.send_transaction(signed_transaction)
```