[DEX] ICX Supply Liquidity Fix (#65)
Switches from modify order to cancel old order, replace new order.
When cancelling the order that was about to be transacted next, there
was an issue moving the node in the linked list implementation.

Instead, this will cancel the old order and place again at the list head.
From a gas perspective, this is worse, but it uses a much better
exercized path in the linked list implementation.