[DEX] Reward Qualification Threshold (#71)
There are two possible spam attacks available on rewards distribution in the DEX, with the current code base. This is due to the low cost of forcing an iteration in rewards, when rewards cannot be distributed for the day until all iterations are passed.

For brevity, these attacks can be referred to as "Many Transfers" and "Many Holders".

*Many Holders* - By keeping a small amount of dust in LP tokens in each of a very large number of accounts, a user can force a very high number of iterations in the rewards SCORE at a low cost to themself. For a few thousand dollars in gas fees, they could pause reward distribution for days. This commit addresses this by not making an address eligible for rewards unless a minimum threshold of liquidity is passed, checked in the `_get_rewardable_amount()` function.

*Many Transfers* - By keeping an amount above the minimum rewardable amount but transferring it throughout the day, a user can force a large amount of possible iterations. For about $50 in assets plus $500 in gas fees, a user could force on the order of 10**5 iterations. To get around this, this commit makes an address only eligible for iteration if it currently holds above the `_get_rewardable_amount()` amount of balance.

As a result of this update, users should not expect to be paid out unless they hold above the minimum amount to qualify for payments, _at the time rewards are computed_. Additionally, this commit assumes all tokens will have 18 digits of precision. A subsequent update should be added for lower resolution coins without an override in `_get_rewardable_amount()`.

Adds the required method for: https://github.com/balancednetwork/balanced-network-interface/issues/68

Additionally, this has some feedback from PR #83 (credit to @leeheonseung) that I have applied here on the ZERO_SCORE_ADDRESS.