[DEX] Fix combined withdraw on easy remove (#85)
Before, the balance check was improperly doubled (so you had to have enough on deposit
AND in the contract when doing the easy remove. The easy remove method creates 2 txns
which send the assets directly to the user's wallet, instead of leaving them in
the DEX to switch to another pool.