[DEX] Improve iteration on non-18 precision
Addresses #91 and #92.

The default case for getDataBatch can reduce spam attacks in iteration
when there are 18 digits of precision. However, for some tokens that do
not use the standard 18 digits (6 is a popular variant, especially among
stablecoins), this won't work well. Now the DEX tracks token precision
at the time of deposit in a cache, and checks that 25 units of that coin
are on deposit to qualify for reward iteration.