[DEX] Fix withdraw and support min pool size
Fixes https://github.com/balancednetwork/balanced-network-interface/issues/195
Fixes #107

This fixes an issue with withdraw amounts in the DEX, so the proper amount is
credited back to the user. There can still be flooring due to integer precision.

Additionally, this adds the concept of minimum liquidity from uniswap. A user must leave
some minimum amount of liquidity in the pool, which should be a very small amount.