Reduced scope of tokenFallback try...except block.
The try...except is only needed to give a better error message when the supplied data is not a valid json. The other code in the tokenFallback does a lot of stuff, including interscore calls to many other contracts. Any exceptions in all of that should be handled specifically, not lumped into the bad data exception.