Pool Optimization (#110)
Fixes #110
Fixes #129

This changes the `add` method to maximize the liquidity based on the supplied parameters.
Adds eventlogs used in both `add` and `swap` events.
Adds a `getPoolStats()` endpoints that gives an aggregated view of tracked quantities, so
the geometry labs backend can track it.

Now, when calling `add`, any extra tokens are automatically removed back to the user's address
if they set a `withdraw` param to true.

Adds extra validation to the withdraw method, to avoid sending a negative number and diluting other
LPs (critical bugfix).

Creates combined methods to take out loans, stake ICX and create a pool in the jupyter notebook.
These can help speed testing of the sICX/bnUSD pool, and can serve as the guideline for unit tests.