Normalize added LP tokens (#104)
This adds a babylonian sqrt implementation used to calculate normalized LP tokens.

Instead of assigning LPs a fixed amount of tokens as before this approach does:

```
sqrt(token_a_amount * token_b_amount)
```

It additionally requires that the amount of LP tokens is greater than the minimum
liquidity. These amounts are based on the smallest unit of tokens, not scaled up
by their precision (1000 loop, not 1000 icx is the requirement, for example).

Fixes #104