BALN Snapshots (#146)
The DEX should track the current amount of BALN held in particular pools. This adds an API that can be called by the dividends score to access this information. This allows it to credit users for their BALN held, but locked up providing liquidity in markets such as BALN / bnUSD.

This introduces a new DictDB to track the quantity, a snapshotting mechanism almost identical to the total supply one, and an external readonly method to retrieve it.