Minimum Receive (#128) and Revert sicx (#162)
Switch the param in `_swap` from a maxSlippage to minimumReceive.

This param is used to allow users to specify an amount of token they wish
to recieve directly in the frontend. If less than this amount, the txn fails.
This establishes a minimum price in a more user-friendly way than slippage.

The default value is 0. Note that this is a breaking change.

Additionally, any other token sent to the token fallback method calling
`_swap_icx` will be reverted. This helps stop fund loss from incorrect
usage of the dex.