Change _pid to _id (#95) + / in pool (#148)
For consistency in the code, all instances of `_pid` are moved to `_id`.
This better matches the ICON multi-token SCORE standard.

The DEX and related contracts are updated in this diff. It will be a
breaking change for the frontend, so please note before merging this PR.

Additionally, this matches the string names in the backend to those used in the frontend.
This helps to remove confusion for API users.