[security] Revert unused methods (#161)
Currently, nobody is using the LP tokens. This may well change with time, but
currently the 3 methods around token transfer represent extra attack surface,
and no value at the current time.

These methods are:

- transfer
- transferFrom
- setApprovalForAll

The rest of the application is subject to much higher both manual/script based test
coverage, and this is outside of the well-travelled path in this application. Taking
this into account, this commit reverts all transactions to these methods. In a future
update, they can be commented back in via a contract update, but right now the
additional attack surface isn't worth the benefit.