Rewards Distribution Hotfix
Some of the snapshots were in a bad state, and cannot be migrated in a satisfactory way before the next rewards distribution.

This is a temporary fix, replacing the 24h time weighted average price with a direct snapshotting. Because all trading on the DEX halts at this time, these APIs will return consistent values.