Fix erasing of balance in ICX queue
After the first day, balanceOf checks didn't work in the ICX queue because
it checked the balance of msg.sender, instead of the balance of the supplied _owner.
While this worked for senders checking their own balance and anyone checking the
contents in a particular pool, it resulted in wiping balances in the ICX queue after
the initial insertion.