[DEX] Min Liquidity (#204)
This removes the ability for a newly inserted address not to be in the actives
addresses list. It should be paired with a method called by the governance score
to insert old addresses below the minimum into active addresses.

Future inserts are subject to minimum amounts, which are available in the
`getPoolStats()` API.