ICX Swap Eventlog (#150)
This adds an eventlog emitted when the `swap_icx` function is called.
This behaves differently from the other eventlogs, since the total quote
currency in the pool is always zero, and the price comes from the sICX
SCORE.

It can be used to generate volume data for the sICX/ICX pool, and populate
historic data for the charts.