[DEX] Transaction router on the balanced DEX
This introduces transaction routing on the balanced DEX, with up to 10 legs allowable in the path.
Users can send in a payable transaction with ICX or a supported IRC2 token to the DEX, a path,
and optionally a minimum amount to receive. The transaction will be routed through different trades
in the DEX, and reverted if no such path exists or there is insufficient funds returned.