[cont-rewards] Deploy fixes for day1 deployments
Before, the frontend would crash without a few necessary methods in dividends. It required the method for lookup user rewards to return, which it would revert on day 1.

Additionally, it sets the required stakedLp configs in governance when listing new markets, and adds a method to setup the BALN/SICX market.