[draft] Fixes to day context in rewards
Before, the contract was considering the data source platform day, which didn't augument up by one with the introduction of continuous rewards. Instead, this switches the quantity to the actual day.

NOTE: Review with @dbrehmer how we should react if the SCORE isn't called for a day. We may want to preserve the platform day quantity in order to create a prorated emission for each day when computing weight.